commit1

+++ src/Country.java
@@ -25,7 +25,7 @@ public class Country {
         return population;
     }
 
-    public void setPopulation(int population) {
+    public void setPopulation(double population) {
         this.population = population;
     }
 
@@ -33,7 +33,7 @@ public class Country {
         return area;
     }
 
-    public void setArea(int area) {
+    public void setArea(double area) {
         this.area = area;
     }
